fix: adPlaceCategoryToIBTravelTag 함수에서 기존 adPlace에 변경전 정보에 태그가 없을 경우 예외처리

+++ src/routes/myPage/myPage.ts
@@ -58,7 +58,7 @@ export const adPlaceCategoryToIBTravelTag = async (param: {
       },
     },
   });
-  if (!isNil(tx) && !isNil(adPlaceId)) {
+  if (!isNil(tx) && !isNil(adPlaceId) && !isEmpty(prevTags)) {
     const a = Prisma.sql`delete from _AdPlaceToIBTravelTag where A = ${adPlaceId} and B in (${Prisma.join(
       prevTags.map(v => v.id),
     )});`;
